Move LeafNode into correct package
