+++ src/main/java/com/meti/evaluate/tokenizer/VariableNodeTokenizer.java
@@ -1,5 +1,6 @@
 package com.meti.evaluate.tokenizer;
 
+import com.meti.render.LeafNode;
 import com.meti.util.Monad;
 import com.meti.render.NodeGroup;
 import com.meti.content.Content;
